feat: Move NoiseModel generation to a set of helper functions and call inside AwsDevice
